Add Python 3 support
1. We can no longer pick the ob_type directly from the
struct, so code like ob->ob_type stops working.
We should replace this with Py_TYPE(ob). The Py_TYPE macro is
supported starting with Python 2.6.

2. Python 3 has now separate types for String, which doesn't
exist anymore. The new types are PyBytes and PyUnicode. We
will replace all usages of PyString with PyUnicode.

3. The family of functions to initialize modules, such as
Py_InitModule3 are gone. Instead, we should use PyModule_Create.

4. Zero_PyTypeObject tries to use the field tp_compare which is
specific to Python 2. On Python 3 is called tp_reserved (<= Python 3.4)
and tp_as_sync ( > Python 3.4).

5. The function which initializes the module changed in Python 3.
In Python 2 was:
  PyMODINIT_FUNC init<yourmodulename>(void)
and in Python 3 is:
  PyMODINIT_FUNC PyInit_<yourmodulename>(void)
We will use a macro to replace the init function name.

6. Using only METH_KEYWORDS will land you with a SystemError that
complains about METH_OLDARGS on Python 3. Note that we can use
METH_VARARGS | METH_KEYWORDS which is equivalent to METH_KEYWORDS
and works on Python 2 and Python 3.