Fix strerror_r usage in socket_wrapper.cpp
Inside the socket_wrapper.cpp file there is a call to strerror_r
which raises an error. As a workaround, the fpermissive flag can
be used which will allow the code to compile even if it doesn't
conform the language rules.

The fpermissive compilation flag can be avoided by removing the
definitions from the unnamed namespace and including server.hpp
in socket_wrapper.cpp.

This patch also changes the usage of strerror_r to take into
account the XSI-compliant version.