Change architecture detection flag from -p to -m
In order to detect the architecture type, uname is being used.
It is given '-p' argument which according to the man page,
it will print the processor type or "unknown" in case it is not
known.

This will print "unknown" on Debian 7, while on Ubuntu platforms
it works correctly printing the architecture type.

Changing the flag to -m will make uname always print the architecture
type on all the systems.