Fix setting 0 value for real numbers
For real numbers, the given value is checked
to be between DBL_MIN and DBL_MAX.

However, DBL_MIN is actually the smallest absolute
value representable, which is greater than 0.0 and
this will be interpreted as out of range.

Adding a case to check for 0.0 before checking the
actual limits fixes the issue.