Fix memory leaks Python >= 3
PyUnicode_AsEncodedString returns a new reference on which we need
to call Py_DECREF after the call to PyBytes_AsString.

To fix this, the macro for PyString_AsString has been removed.
All the usages of that function have been replaced with the correct
one that also frees the object created.